Add MQTT controllable P1 GPIO
P1 is the LED strip output on the power module. By connecting it to an
external relay a secondary high current switch can easily be added. In
my project I intend to control it from a Node-RED flow so I make P1
controllable over radio similar to the power module relay in the
twr-node-generic firmware.